requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioredis==1.3.1
aiosignal==1.4.0
aiosmtplib==4.0.1
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asgiref==3.9.1
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.3.0
billiard==4.2.1
black==25.9.0
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.11.0
cryptography==45.0.5
elastic-transport==9.1.0
elasticsearch==9.1.0
fastapi==0.116.1
fastapi-cache==0.1.0
fastapi-cache2==0.2.2
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
hiredis==3.2.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.3.0
isort==6.0.1
Jinja2==3.1.6
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.3
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pendulum==3.1.0
platformdirs==4.4.0
pluggy==1.6.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytokens==0.1.10
redis==6.2.0
six==1.17.0
sniffio==1.3.1
sqladmin==0.21.0
SQLAlchemy==2.0.41
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
WTForms==3.1.2
yarl==1.20.1
